pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "quantumthink"
version = "0.1.0"
description = "Production-ready starter repository for an AI agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "QuantumThink" }
]
dependencies = [
  "PyYAML>=6.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]
openai = [
  "openai>=1.40.0"
]

[project.scripts]
quantumthink = "quantumthink.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
